requirements.txt
flask
flask-sqlalchemy
flask-migrate
psycopg2-binary
dotenv
gunicorn
marshmallow
pytz